moved all  quandl api requests to client
